deploy: Copy changed files only
* Add --force option to deploy

+++ tasks/deploy.py
@@ -1,13 +1,16 @@
 import shutil
 from invoke import task
-from tasks.build import build
+from .build import build
+from .sync import sync_dirs
 
 # TODO allow specifying target bf1942 install
 @task(build)
-def deploy(c):
+def deploy(c, force=False):
     deploy_path = c.bf1942.path / 'Mods' / c.mod.name
 
-    if deploy_path.exists():
-        shutil.rmtree(deploy_path)
-    shutil.copytree(c.pack_path, deploy_path)
+    if force:
+        if deploy_path.exists():
+            shutil.rmtree(deploy_path)
+        deploy_path.mkdir(parents=True)
 
+    sync_dirs(c.pack_path, deploy_path)
